+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to **ValueStringBuilder** will be documented in this file. T
 
 ## [Unreleased]
 
+### Fixed
+
+- `Write` did not return the correct amount of added characters
+
 ## [0.6.0] - 2022-12-28
 
 ### Added
@@ -14,7 +18,7 @@ All notable changes to **ValueStringBuilder** will be documented in this file. T
 
 ## [0.5.3] - 2022-12-28
 
-\### Changed
+### Changed
 
 -   Smaller internal refactorings